fixed signedness
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5840 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/powermgmt.c
@@ -209,7 +209,7 @@ int voltage_to_percent(int voltage, const short* table)
 void battery_level_update(void)
 {
     int level = 0;
-    int c = 0;
+    unsigned short c = 0;
     int i;
 #if BATTERY_TYPES_COUNT == 1 /* single type */
     const int battery_type = 0;
